backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
dill==0.3.9
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
packaging==26.0
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.2
pycparser==3.0
pydantic==2.11.7
pydantic_core==2.33.2
pydrantic==0.0.3
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.35.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
